undo previous new files and use existing encoding_utf header
