+++ gfx/drivers_font/gl_raster_font.c
@@ -17,7 +17,7 @@
 #include "../common/gl_common.h"
 #include "../font_driver.h"
 #include "../video_shader_driver.h"
-#include <string/utf8_util.h>
+#include <encodings/utf.h>
 
 /* TODO: Move viewport side effects to the caller: it's a source of bugs. */
 
@@ -35,14 +35,6 @@
 } while(0)
 
 #define MAX_MSG_LEN_CHUNK 64
-
-#ifdef HAVE_UTF8
-#define string_walk utf8_walk
-#define string_len utf8_strlen
-#else
-#define string_walk utf8_walkbyte
-#define string_len strlen
-#endif
 
 typedef struct
 {
@@ -307,7 +299,7 @@ static void gl_raster_font_render_line(
       {
          int off_x, off_y, tex_x, tex_y, width, height;
          const struct font_glyph *glyph =
-            font->font_driver->get_glyph(font->font_data, string_walk(&msg));
+            font->font_driver->get_glyph(font->font_data, utf8_walk(&msg));
 
          if (!glyph) /* Do something smarter here ... */
             glyph = font->font_driver->get_glyph(font->font_data, '?');
@@ -369,7 +361,7 @@ static void gl_raster_font_render_message(
    /* If the font height is not supported just draw as usual */
    if (!font->font_driver->get_line_height)
    {
-      gl_raster_font_render_line(font, msg, string_len(msg),
+      gl_raster_font_render_line(font, msg, utf8len(msg),
             scale, color, pos_x, pos_y, text_align);
       return;
    }
@@ -393,7 +385,7 @@ static void gl_raster_font_render_message(
       }
       else
       {
-         unsigned msg_len = string_len(msg);
+         unsigned msg_len = utf8len(msg);
          gl_raster_font_render_line(font, msg, msg_len, scale, color, pos_x,
                pos_y - (float)lines*line_height, text_align);
          break;
