+++ libretro-common/encodings/encoding_utf.c
@@ -208,3 +208,31 @@ size_t utf8len(const char *string)
    return strlen(string);
 #endif
 }
+
+uint8_t utf8_walkbyte(const char **string)
+{
+   return *((*string)++);
+}
+
+/* Does not validate the input, returns garbage if it's not UTF-8. */
+uint32_t utf8_walk(const char **string)
+{
+   uint8_t first = utf8_walkbyte(string);
+   uint32_t ret;
+   
+   if (first<128)
+      return first;
+   
+   ret = 0;
+   ret = (ret<<6) | (utf8_walkbyte(string)    & 0x3F);
+   if (first >= 0xE0)
+      ret = (ret<<6) | (utf8_walkbyte(string) & 0x3F);
+   if (first >= 0xF0)
+      ret = (ret<<6) | (utf8_walkbyte(string) & 0x3F);
+   
+   if (first >= 0xF0)
+      return ret | (first&31)<<18;
+   if (first >= 0xE0)
+      return ret | (first&15)<<12;
+   return ret | (first&7)<<6;
+}
